deep-observer/ai-brain/requirements.txt
clickhouse-connect==0.8.18
psycopg[binary]==3.2.13
python-dotenv==1.0.1
requests==2.32.3
